refactor: add CameraFrames
