+++ src/app/Scenes/ReverseImage/Components/FrameIndicator.tsx
