+++ src/app/Scenes/ReverseImage/Components/FrameIndicators.tsx
